ElasticSearchBlogProject/requirements.txt
albumentations==1.3.1
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
bcrypt==4.2.0
blinker==1.8.2
boto3==1.28.40
botocore==1.31.40
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
constantly==15.1.0
contourpy==1.2.0
cryptography==41.0.4
cycler==0.12.1
Cython==3.0.7
daphne==4.0.0
decorator==5.1.1
distlib==0.3.7
Django==4.0
django-ckeditor==6.7.1
django-cors-headers==4.2.0
django-environ==0.4.4
django-extensions==3.2.3
django-filter==23.2
django-js-asset==2.2.0
django-storages==1.14
django-treebeard==4.7
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
dlib==19.24.2
drf-yasg==1.21.7
easydict==1.11
exceptiongroup==1.2.0
executing==1.2.0
filelock==3.13.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
flatbuffers==23.5.26
fonttools==4.47.0
googlemaps==4.10.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
imageio==2.33.1
importlib-resources==6.1.1
importlib_metadata==8.2.0
incremental==22.10.0
inflection==0.5.1
insightface==0.7.3
ipython==8.12.2
itsdangerous==2.2.0
jedi==0.19.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mpmath==1.3.0
msgpack==1.0.5
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
onnx==1.15.0
onnxruntime==1.16.3
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.0.0
pooch==1.8.0
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==4.25.1
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.16.1
PyJWT==2.9.0
PyMatting==1.1.12
pyOpenSSL==23.2.0
pyotp==2.9.0
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3
PyYAML==6.0.1
qudida==0.0.4
razorpay==1.4.1
redis==5.0.0
referencing==0.32.0
rembg==2.0.53
replicate==0.22.0
requests==2.31.0
rpds-py==0.15.2
s3transfer==0.6.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.32
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.12.9
tqdm==4.66.1
traitlets==5.9.0
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.24.7
wcwidth==0.2.6
Werkzeug==3.0.3
zipp==3.17.0
zope.interface==6.0
